+++ command/agent/http.go
@@ -175,8 +175,10 @@ func (s *HTTPServer) wrap(handler func(resp http.ResponseWriter, req *http.Reque
 		}
 
 		prettyPrint := false
-		if _, ok := req.URL.Query()["pretty"]; ok {
-			prettyPrint = true
+		if v, ok := req.URL.Query()["pretty"]; ok {
+			if len(v) > 0 && (len(v[0]) == 0 || v[0] != "0") {
+				prettyPrint = true
+			}
 		}
 
 		// Write out the JSON object
@@ -184,6 +186,9 @@ func (s *HTTPServer) wrap(handler func(resp http.ResponseWriter, req *http.Reque
 			var buf []byte
 			if prettyPrint {
 				buf, err = json.MarshalIndent(obj, "", "    ")
+				if err == nil {
+					buf = append(buf, "\n"...)
+				}
 			} else {
 				buf, err = json.Marshal(obj)
 			}