+++ command/agent/http_test.go
@@ -145,14 +145,18 @@ func TestContentTypeIsJSON(t *testing.T) {
 }
 
 func TestPrettyPrint(t *testing.T) {
-	testPrettyPrint("pretty=1", t)
+	testPrettyPrint("pretty=1", true, t)
+}
+
+func TestPrettyPrintOff(t *testing.T) {
+	testPrettyPrint("pretty=0", false, t)
 }
 
 func TestPrettyPrintBare(t *testing.T) {
-	testPrettyPrint("pretty", t)
+	testPrettyPrint("pretty", true, t)
 }
 
-func testPrettyPrint(pretty string, t *testing.T) {
+func testPrettyPrint(pretty string, prettyFmt bool, t *testing.T) {
 	s := makeHTTPServer(t, nil)
 	defer s.Cleanup()
 
@@ -167,14 +171,20 @@ func testPrettyPrint(pretty string, t *testing.T) {
 	req, _ := http.NewRequest("GET", urlStr, nil)
 	s.Server.wrap(handler)(resp, req)
 
-	expected, _ := json.MarshalIndent(r, "", "    ")
+	var expected []byte
+	if prettyFmt {
+		expected, _ = json.MarshalIndent(r, "", "    ")
+		expected = append(expected, "\n"...)
+	} else {
+		expected, _ = json.Marshal(r)
+	}
 	actual, err := ioutil.ReadAll(resp.Body)
 	if err != nil {
 		t.Fatalf("err: %s", err)
 	}
 
 	if !bytes.Equal(expected, actual) {
-		t.Fatalf("bad: %q", string(actual))
+		t.Fatalf("bad:\nexpected:\t%q\nactual:\t\t%q", string(expected), string(actual))
 	}
 }